some new fragments and teacher and teacher-detail

+++ src/main/java/org/sb/eduhome2/repositories/BlogRepository.java
@@ -1,7 +1,8 @@
 package org.sb.eduhome2.repositories;
 
+import jakarta.persistence.criteria.CriteriaBuilder;
 import org.sb.eduhome2.models.Blog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlogRepository  extends JpaRepository<Blog, Long> {
+public interface BlogRepository  extends JpaRepository<Blog, Integer> {
 }
